# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация к тестовому заданию",
        "contact": {
            "email": "dmitriylasarev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/cars": {
            "get": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Получить список автомобилей",
                "description": "Возвращает список автомобилей",
                "operationId": "getCarsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/drivers": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Получить список пользователей",
                "description": "Возвращает список пользователей",
                "operationId": "getDriversList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriversResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Получить информацию об автомобиле",
                "description": "Возвращает информацию об автомобиле по ID",
                "operationId": "getCarById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarsResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/bookCar": {
            "post": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Бронируем автомобиль",
                "description": "Бронируем автомобиль для указанного пользователя",
                "operationId": "bookCar",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "query",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "driver_id",
                        "in": "query",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarsResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/releaseCar": {
            "post": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Освобождаем автомобиль",
                "description": "Освобождаем указанный автомобиль",
                "operationId": "releaseCar",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "query",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarsResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/drivers/{id}": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Получить информацию о пользователе",
                "description": "Возвращает информацию об указанном пользователе",
                "operationId": "getDriverById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriversResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CarsResource": {
                "title": "CarsResource",
                "description": "Cars resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Car"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CarsResource"
                }
            },
            "DriversResource": {
                "title": "DriversResource",
                "description": "Drivers resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Driver"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriversResource"
                }
            },
            "Car": {
                "title": "Car",
                "description": "Модель автомобиля",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "model": {
                        "title": "model",
                        "description": "Модель автомобиля",
                        "type": "string",
                        "example": "Mersedes Benz"
                    },
                    "sign": {
                        "title": "sign",
                        "description": "Номерной знак автомобиля",
                        "type": "string",
                        "example": "S123ZX"
                    },
                    "occuped_by": {
                        "title": "occuped_by",
                        "description": "ID пользователя, забронировавшего авто",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Car"
                }
            },
            "Driver": {
                "title": "Driver",
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван Иванов"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Driver"
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Автомобили",
            "description": "API для работы с автомобилями"
        },
        {
            "name": "Пользователи",
            "description": "API для работы с пользователями"
        }
    ]
}